# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A full-stack development basic platform using gin+vue for extremely fast development",
        "title": "Gin-Vue-Admin Swagger API Interface Document",
        "contact": {},
        "version": "v2.5.9"
    },
    "paths": {
...
        "/authorityBtn/canRemoveAuthorityBtn": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthorityBtn"
                ],
                "summary": "Set permission button",
                "responses": {
                    "200": {
                        "description": "Delete successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authorityBtn/getAuthorityBtn": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthorityBtn"
                ],
                "summary": "Get permission button",
                "parameters": [
                    {
                        "description": "menu id, role id, selected button id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysAuthorityBtnReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return list successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SysAuthorityBtnRes"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/authorityBtn/setAuthorityBtn": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AuthorityBtn"
                ],
                "summary": "Set permission button",
                "parameters": [
                    {
                        "description": "menu id, role id, selected button id",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.SysAuthorityBtnReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return list successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/email/emailTest": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Send test email",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"Sent successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/email/sendEmail": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Send email",
                "parameters": [
                    {
                        "description": "Required parameters for sending emails",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/response.Email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"Sent successfully\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.SysAuthorityBtnReq": {
            "type": "object",
            "properties": {
                "authorityId": {
                    "type": "integer"
                },
                "menuID": {
                    "type": "integer"
                },
                "selected": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "response.Email": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "email content",
                    "type": "string"
                },
                "subject": {
                    "description": "Email title",
                    "type": "string"
                },
                "to": {
                    "description": "To whom the email is sent",
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.SysAuthorityBtnRes": {
            "type": "object",
            "properties": {
                "selected": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}